Axed config vars, most of alchemist, refactored argparser
