+++ horde-scribe-bridge.cmd
@@ -1,3 +1,3 @@
 @echo off
 cd /d %~dp0
-call runtime python -s bridge_scribe.py %*
+call runtime python -s horde-scribe-bridge.sh %*
