+++ horde-scribe-bridge.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-./runtime.sh python -s bridge_scribe.py "$@"
+./runtime.sh python -s horde-scribe-bridge.py "$@"
