--- worker/__init__.py
@@ -1,4 +1,0 @@
-from worker.utils.switch import Switch
-
-disable_xformers = Switch()
-disable_download_progress = Switch()
